used foundry lint

+++ src/interfaces/compound/IComet.sol
@@ -2,9 +2,9 @@
 pragma solidity 0.8.33;
 
 interface IComet {
-    function supply(address asset, uint amount) external;
+    function supply(address asset, uint256 amount) external;
 
-    function withdraw(address asset, uint amount) external;
+    function withdraw(address asset, uint256 amount) external;
 
     function balanceOf(address owner) external view returns (uint256);
 }
